nuevo contexto

+++ Models/ProcesoTitulacion.cs
@@ -9,37 +9,37 @@ public partial class ProcesoTitulacion
 
     public string NoControl { get; set; } = null!;
 
-    public ulong Paso1 { get; set; }
+    public int Paso1 { get; set; }
 
-    public ulong Scni { get; set; }
+    public int Scni { get; set; }
 
-    public ulong Cni { get; set; }
+    public int Cni { get; set; }
 
-    public ulong Cl { get; set; }
+    public int Cl { get; set; }
 
-    public ulong Caii { get; set; }
+    public int Caii { get; set; }
 
-    public ulong Rp { get; set; }
+    public int Rp { get; set; }
 
-    public ulong Rps { get; set; }
+    public int Rps { get; set; }
 
-    public ulong St { get; set; }
+    public int St { get; set; }
 
-    public ulong Pro { get; set; }
+    public int Pro { get; set; }
 
-    public ulong Sl { get; set; }
+    public int Sl { get; set; }
 
-    public ulong Lp { get; set; }
+    public int Lp { get; set; }
 
-    public ulong Asnc { get; set; }
+    public int Asnc { get; set; }
 
-    public ulong Oi { get; set; }
+    public int Oi { get; set; }
 
-    public ulong Curp { get; set; }
+    public int Curp { get; set; }
 
-    public ulong Rfc { get; set; }
+    public int Rfc { get; set; }
 
-    public ulong Cb { get; set; }
+    public int Cb { get; set; }
 
     public virtual InfoPersonal NoControlNavigation { get; set; } = null!;
 }
